requirements.txt
asgiref==3.7.2
Django==4.2.11
djangorestframework==3.14.0
model-bakery==1.17.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
django-virtual-models==0.2.0
